fixed in subject

+++ database/migrations/2019_08_07_210721_create_school_program_subject_table.php
@@ -19,7 +19,7 @@ class CreateSchoolProgramSubjectTable extends Migration
             $table->unsignedBigInteger('subject_id');
             $table->string('type',2)
                 ->nullable();
-            $table->unsignedBigInteger('subject_group')
+            $table->integer('subject_group')
                 ->nullable();
             $table->foreign('school_program_id')
                 ->references('id')
